pyproject.toml
[project]
name = "kavzitrader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "plotly>=6.5.1",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.44.0",
    "openai>=1.50.0",
    "jinja2>=3.1.0",
    "pytest-env>=1.2.0",
    "python-binance>=1.0.34",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "redis>=7.1.0",
    "websockets>=15.0.1",
    "aiofiles>=25.1.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "types-redis>=4.6.0.20241004",
    "types-pyyaml>=6.0.12.20250915",
    "types-aiofiles>=24.1.0.20241019",
]


[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)s] %(filename)s:%(lineno)s %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
asyncio_default_test_loop_scope = "function"

testpaths = [
    "tests",
]


[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D",
    "TD",
    "FIX",
    "S311",
    "S603",
    "S608", # TODO: Remove this
    "PLR0913",
    "PGH003",
    "PTH123",
    "PTH118",
    "PTH119",
    "TRY003",
    "EM101",
    "EM102",
    "FBT003",
    "FBT001",
    "TC001",
    "TC002",
    "G004",
    "FBT002",
    "COM812",
    "TC003",
    "PLC0415",
    "ASYNC110",
    "PLR0915",
    "C901",
    "PLW0108",
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query", "fastapi.Path", "fastapi.Body"]

[tool.ruff.lint.per-file-ignores]
"**/{tests,docs,tools}/*" = ["S101", "S106", "S108", "SLF001", "PLR0913", "TRY301", "PT011", "ANN", "ARG", "PTH", "PERF", "PLR2004", "S110", "BLE001"]
"**/{scripts,examples}/*" = ["T201", "S607"]
